Update components/P3WidgetContainer.php
Content can be used as data for the used Widget, because the content is rendered in between 
the beginWidget() (calls function init() of the widget class) 
and endWidget() (calls function run() of the widget class) functions. 

Or if in the content the tag '{WIDGET_CONTENT}' is used, the Widget is rendered within the content (usually text / pictures ) with only the set parameters.